Add files via upload
The probability evaluation of  Start-up Success/Failure 
When we investigate the statistics of startups, one of the hard-hitting facts is 90% of startups are prone to fail for various reasons. That reason might be poles apart, including unsatisfactory administration, insufficient capital, etc. The concept behind the aim here is to construct a predictive model for a company start-up based on various key factors involving the company's growth.